refactor: SignUpPage.jsx에서 SignUp.jsx로 코드 이동
